+++ R/run_model.R
@@ -38,6 +38,10 @@
 #'
 #' @param mindiff A `numeric` strictly positive value indicating the
 #' stopping criterion for adjusting the Doubly Constrained Model (see Details).
+#' 
+#' @param eps A strictly positive `numeric` value used to replace any zero
+#' values in the marginals of the Doubly Constrained Model, in order to avoid
+#' numerical issues during the IPF procedure (see Details).
 #'
 #' @param check_names A `boolean` indicating whether the location IDs used as 
 #' matrix rownames and colnames should be checked for consistency 
@@ -69,7 +73,9 @@
 #' default) and `mindiff` (0.01 by default) can be used to tune the model.
 #' `mindiff` is the minimal tolerated relative error between the
 #' simulated and observed marginals. `maxiter` ensures that the algorithm stops
-#'  even if it has not converged toward the `mindiff` wanted value.
+#'  even if it has not converged toward the `mindiff` wanted value. The argument
+#' `eps` (1e-6 by default) can be used to replace zero values in the marginals
+#' in order to avoid numerical issues during the IPF procedure.
 #'
 #' By default, when `average = FALSE`, `nbrep` matrices are generated from
 #' `proba` with multinomial random draws that will take different forms
@@ -118,14 +124,17 @@
 #' Oi <- as.numeric(mass[, 2])
 #' Dj <- as.numeric(mass[, 3])
 #'
-#' res <- run_model(
-#'   proba = proba,
-#'   model = "DCM", nb_trips = NULL, out_trips = Oi, in_trips = Dj,
-#'   average = FALSE, nbrep = 3, maxiter = 50, mindiff = 0.01,
-#'   check_names = FALSE
-#' )
-#'
-#' # print(res)
+#' res <- run_model(proba = proba,
+#'                  model = "DCM", 
+#'                  nb_trips = NULL, 
+#'                  out_trips = Oi, 
+#'                  in_trips = Dj,
+#'                  average = FALSE, 
+#'                  nbrep = 3, 
+#'                  maxiter = 50, 
+#'                  mindiff = 0.01,
+#'                  eps = 0.000001,
+#'                  check_names = FALSE)
 #'
 #' @export
 run_model <- function(proba,
@@ -137,6 +146,7 @@ run_model <- function(proba,
                       nbrep = 3,
                       maxiter = 50,
                       mindiff = 0.01,
+                      eps = 0.000001,
                       check_names = FALSE) {
   
   # Option (disabling scientific notation)
@@ -195,9 +205,11 @@ run_model <- function(proba,
   if (model == "DCM") {
     controls(args = maxiter, type = "strict_positive_integer")
     controls(args = mindiff, type = "strict_positive_numeric")
+    controls(args = eps, type = "strict_positive_numeric")
   } else {
     maxiter <- "50"
     mindiff <- "0.01"
+    eps <- "0.000001"
   }
 
   if (model == "UM") {
@@ -341,6 +353,7 @@ run_model <- function(proba,
   }
   maxiterDCM <- maxiter
   minratioDCM <- mindiff
+  epsDCM <- eps
 
   outputs <- list()
   Args <- c("Model", "#Replications")
@@ -352,7 +365,7 @@ run_model <- function(proba,
 
   args <- paste0(
     wdin, " ", wdout, " ", model, " ", nbrep, " ", multi, " ",
-    maxiterDCM, " ", minratioDCM
+    maxiterDCM, " ", minratioDCM, " ", epsDCM
   )
 
   cmd <- paste0("java -jar ", wdjar, "TDM.jar ", args)