+++ NEWS.md
@@ -3,9 +3,17 @@
 This is a list of changes made in the development/GitHub version of the package  
 between TDLM 1.1.3 (CRAN release 2025-09-03) and the next CRAN release.
 
+* Updated `vignette`.
+
+* Added the `eps` argument to `run_law_model()` and `run_model()` allowing to
+choose the value used to replace zero values in the marginals of the Doubly 
+Constrained Model (in order to avoid numerical issues during the IPF procedure).
+The value 0.01 was hard-coded in the function before.
+
 * Corrected an issue with `DCM` (`run_law_model()` and `run_model()`) and the 
 way it handles zeroes in pij. It was particularly problematic with 
-`law = "Unif"`.
+`law = "Unif"` (
+https://github.com/RTDLM/TDLM/blob/master/java/TDLM.java#L505).
 
 # TDLM 1.1.3